extract styled components into common file
