+++ tutor/src/screens/assignment-review/reading-scores.js
@@ -1,7 +1,6 @@
-import { React, PropTypes, styled, observer, css } from 'vendor';
-import { StickyTable, Row, Cell as TableCell } from 'react-sticky-table';
+import { React, PropTypes, styled, observer } from 'vendor';
+import { StickyTable, Row } from 'react-sticky-table';
 import { OverlayTrigger, Popover } from 'react-bootstrap';
-import { Icon } from 'shared';
 import LoadingScreen from 'shared/components/loading-animation';
 import { colors } from 'theme';
 import S from '../../helpers/string';
@@ -15,100 +14,31 @@ import DropQuestions from './drop-questions';
 import PublishScores from './publish-scores';
 import { EIcon } from '../../components/icons/extension';
 
+import {
+  StyledStickyTable, Cell,
+  CellContents as BasicCellContents,
+  Heading, HeadingTop, HeadingMiddle, HeadingBottom,
+  ColumnHeading as BasicColumnHeading,
+  SplitCell, ColumnFooter, Total,
+  LateWork, DefinitionsWrapper,
+  ControlsWrapper, ControlGroup,
+  OrderIcon, NameWrapper,
+} from './table-elements';
+
 // https://projects.invisionapp.com/d/main#/console/18937568/401942280/preview
 
-const StyledStickyTable = styled(StickyTable)`
-  margin: 2.2rem 0 1.4rem;
-
-  .sticky-table-row:last-child .sticky-table-cell {
-    border-bottom: 1px solid ${colors.neutral.pale};
-  }
-`;
-
-const Cell = styled(TableCell)`
-  padding: 0;
-  border-width: 1px;
-  border-color: transparent;
-  border-left: 1px solid ${colors.neutral.pale};
-  cursor: ${props => props.onClick || props.clickable ? 'pointer' : 'inherit'};
-  &:last-child {
-    border-right: 1px solid ${colors.neutral.pale};
-  }
-  ${props => props.striped && css`
-    background: ${colors.neutral.lighter};
-  `}
-  && {
-    ${props => props.isTrouble && isTroubleCSS}
-  }
-`;
-
-const centeredCSS = css`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const paddingCSS = css`
-  padding: 1.2rem 1.6rem;
-`;
-
-const CellContents = styled.div`
-  ${centeredCSS}
-  > * { width: 80px; }
-  > *:first-child {
-    width: 16rem;
-  }
+const CellContents = styled(BasicCellContents)`
   .data-heading {
     height: 110px;
   }
 `;
 
-const Heading = styled.div`
-  ${props => !props.first && centeredCSS}
-  ${props => props.first && !props.noBorder && css`
-    border-right: 1px solid ${colors.neutral.pale};
-  `}
-  ${paddingCSS}
+const ExtIcon = styled(EIcon)`
+  display: inline-block;
 `;
 
-const HeadingTop = styled.div`
-  ${paddingCSS}
-  padding-top: 1.2rem;
-  align-self: stretch;
-  font-weight: bold;
-  cursor: pointer;
-`;
-
-const HeadingMiddle = styled.div`
-  ${paddingCSS}
-  align-self: stretch;
-  padding-top: 0;
-  font-size: 1rem;
-  color: ${colors.neutral.thin};
-`;
-
-const HeadingBottom = styled.div`
-  ${paddingCSS}
-  align-self: stretch;
-  font-size: 1.2rem;
-  line-height: 1.4rem;
-  color: ${colors.neutral.thin};
-  background: #fff;
-  position: relative;
-`;
-
-const ColumnHeading = styled.div`
+const ColumnHeading = styled(BasicColumnHeading)`
   background: ${props => props.variant === 'q' ? colors.templates.reading.background : colors.neutral.lighter};
-  border-top: 0.4rem solid ${props => props.variant === 'q' ? colors.templates.reading.border : colors.neutral.std};
-  cursor: ${props => props.onClick || props.clickable ? 'pointer' : 'inherit'};
-  &:not(:last-child) {
-    border-right: 1px solid ${colors.neutral.pale};
-  }
-  > * {
-    ${props => !props.first && css`
-      ${centeredCSS}
-    `}
-  }
   .middle-data-heading {
     padding-top: 15px;
   }
@@ -117,80 +47,6 @@ const ColumnHeading = styled.div`
   }
 `;
 
-const ColumnFooter = styled.div`
-
-  &:not(:last-child) {
-    border-right: 1px solid ${colors.neutral.pale};
-  }
-  > * {
-    /* TODO change to 1.6rem across the board */
-    font-size: 1.4rem;
-    ${props => !props.first && css`
-      ${centeredCSS}
-    `}
-  }
-`;
-
-const SplitCell = styled.div`
-  ${centeredCSS}
-  flex: 1.0;
-  font-size: 1.4rem;
-  ${props => props.variant != 'active' && `color: ${colors.link};`}
-  ${props => props.variant == 'divider' && `color: ${colors.neutral.pale};`}
-  ${props => props.variant != 'divider' && 'cursor: pointer;'}
-`;
-
-const LateWork = styled.div`
-  padding: 0;
-  ${centeredCSS}
-  align-self: stretch;
-  position: relative;
-
-  .extension-icon {
-    position: absolute;
-    right: 1rem;
-  }
-`;
-
-const Total = styled.div`
-  padding: 0;
-  align-self: stretch;
-  border-right: 1px solid  ${colors.neutral.pale};
-  ${centeredCSS}
-`;
-
-const isTroubleCSS = css`
-  background-color: ${colors.states.trouble};
-  border-top: 1px solid ${colors.danger};
-  border-bottom: 1px solid ${colors.danger} !important;
-`;
-
-const DefinitionsWrapper = styled.div`
-  margin: 1.4rem 0;
-  ol {
-    margin-top: 8px;
-    padding-left: 15px;
-    
-    li:not(:first-child) {
-      margin-top: 8px;
-      .extension-icon {
-        display: inline-block;
-      }
-    }
-  }
-`;
-
-const OrderIcon = styled(Icon)`
-  &&.btn {
-    transition: none;
-    font-size: 1.2rem;
-    line-height: 1.2rem;
-
-    &:hover, &:focus {
-      box-shadow: none;
-    }
-  }
-`;
 
 const popover = (gradingTemplate) => (
   <Popover className="scores-popover">
@@ -287,7 +143,6 @@ const StudentColumnHeader = observer(({ ux }) => (
   </Cell>
 ));
 
-
 const StudentCell = observer(({ ux, student, striped }) => {
   const countData = ux.getReadingCountData(student);
   return (
@@ -295,7 +150,7 @@ const StudentCell = observer(({ ux, student, striped }) => {
       <Cell striped={striped}>
         <CellContents>
 
-          <Heading first={true}>
+          <NameWrapper first>
             <TutorLink
               to="viewTask"
               params={{
@@ -305,7 +160,7 @@ const StudentCell = observer(({ ux, student, striped }) => {
             >
               {ux.reverseNameOrder ? student.reversedName : student.name}
             </TutorLink>
-          </Heading>
+          </NameWrapper>
 
           <Total>
             {ux.displayTotalInPercent ?
@@ -372,35 +227,6 @@ const AverageScoreHeader = observer(({ ux }) => (
 ));
 
 
-const ControlsWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 3rem;
-`;
-
-const ControlGroup = styled.div`
-  display: flex;
-
-  .input-group {
-    width: 25.6rem;
-
-    input {
-      height: 3.8rem;
-    }
-  }
-
-  && {
-    > *:not(:first-child), > .btn + .btn {
-      margin-left: 1.6rem;
-    }
-    .btn:not(.btn-icon) {
-      height: 4rem;
-      min-width: 17rem;
-    }
-  }
-`;
-
 const TableHeader = observer(({ ux }) => {
   return (
     <ControlsWrapper>
@@ -449,7 +275,7 @@ const ReadingScores = observer(({ ux }) => {
         <strong>NOTE</strong>
         <ol>
           <li>The late penalty is applied only to the points earned after the due date. <a href="https://openstax.org/blog/new-openstax-tutor-scoring-feature" target="_blank">Learn more</a></li>
-          <li>Students who’ve been granted an extension are denoted with <EIcon /></li>
+          <li>Students who’ve been granted an extension are denoted with <ExtIcon /></li>
         </ol>
       </DefinitionsWrapper>
     </>
